Move some Windows Winword NVDAObject stuff into IAccessible WinWord NVDAObject.
Use UIA with Microsoft word if NVDA connot inject, or NVDA is specifically configured to do so.

Choose WordDocument with IAccessible

Outlook appModule: grab WordDocument code from IAccessible rather than Window
